FIX indexing, get, count after slicing
Previously, calling get or count, or indexing a query set after it has
been sliced would ignore the slice. This has been fixed by keeping track
of slices, not just for the immediate query itself but future indexing
and slicing operations. Count method has also been fixed to respect the
slicing parameters, and return the number that is actually accessible
with the current query parameters. It is also now possible to slice a
query set that has already been sliced.

rref #5441